% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counterfactual.R
\name{cf_plot}
\alias{cf_plot}
\title{Plot original vs. counterfactual survival curves}
\usage{
cf_plot(cf_out, main = "Counterfactual")
}
\arguments{
\item{cf_out}{A list returned by cf_compare_one().}

\item{main}{Character; plot title.}
}
\value{
A base R plot displaying survival curves.
}
\description{
Simple visualization comparing the original and counterfactual survival curves.
}
\examples{
# cf_plot(cf_result)
}
