% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_optimizer_from_str}
\alias{get_optimizer_from_str}
\title{Map string to torch optimizer function}
\usage{
get_optimizer_from_str(opt)
}
\arguments{
\item{opt}{Character. One of "sgd", "adam", or "rmsprop".}
}
\value{
A function or NULL if the string is unrecognized.
}
\description{
Returns the corresponding torch optimizer class given a string.
}
