% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survex_integration.R
\name{make_deepsurv_explainer}
\alias{make_deepsurv_explainer}
\title{Create a survex explainer object for a trained DeepSuRv model}
\usage{
make_deepsurv_explainer(model, data, time_col, event_col)
}
\arguments{
\item{model}{A trained DeepSuRv model object.}

\item{data}{A data.frame containing covariates + survival columns.}

\item{time_col}{Name of the survival time column.}

\item{event_col}{Name of the event indicator column (1 = event).}
}
\value{
A survex explainer object created by survex::explain_survival()
}
\description{
Create a survex explainer object for a trained DeepSuRv model
}
